requirements.txt
Flask
pymongo
gunicorn
pytest
requests
werkzeug
markdown
flask-compress
stripe
openai
chromadb
google-generativeai
sentence-transformers
# Azure SDK packages
azure-cosmos
azure-openai
azure-identity
azure-keyvault-secrets
azure-monitor-opentelemetry
azure-storage-blob
